test_inotify: do not assume watch descriptor values
It is not safe to assume the values returned by add_watch(),
so we add a local helper get_wd() to retrieve them.  This fixes
a problem in TestInotifyTree.test__cycle() where the
wd's for the 'aa' and 'bb' paths are not '2' and '3',
respectively.

A second issue is that Inotify._get_event_names() should
return a set or sorted list to avoid comparison problems,
but that is not addressed here since it could be viewed as
changing the API.

Signed-off-by: Joe Slater <joe.slater@windriver.com>