Eliminate race conditions when tree modified during `InotifyTree()`
If directories are removed from the directory hierarchy passed to
`InotifyTree()` while it is doing the initial scan of directories in
the tree, it can cause a crash, either because it tries to list the
directory and it's no longer there, or it tries to add a watch on the
directory and it's no longer there.

To fix this, we check for `FileNotFoundError` when listing the
directory and `errno == ENOENT` when adding the watch, and log a debug
message and continue when one of them occurs, rather than raising an
exception.

Fixes dsoprea/PyInotify#113